+++ src/qa/boundary/questions.clj
@@ -17,7 +17,10 @@
   duct.database.sql.Boundary
   (create [db nick question]
     (debug "nick" nick "question" question)
-    (sql/insert! (ds db :questions {:nick nick :q question})))
+    (sql/insert! (ds db) :questions {:nick nick :q question}))
+
   (fetch [db n]
     (sql/get-by-id (ds db) :questions n bf))
-  (fetch-all [db]))
+
+  (fetch-all [db]
+    (sql/query (ds db) ["select * from questions"] bf)))